Change `connect_mime_type` to consider public units and same linker scope units.
With this patch we now perform lookup of parsers by MIME type depending
on a given scope so that we return only `public` parsers or parsers with
the same scope.

For that we change the `connect_mime_type` runtime functions to take an
new `scope` parameter which by default uses the `linker_scope` of the
calling code. This means in practical terms that if one calls
`connect_mime_type` on sink without giving a scope only public parsers
as well as parsers in the HLTO file are potentially connected; before we
only connected `public` parsers.